+++ backend/src/app.module.ts
@@ -3,10 +3,14 @@ import { ConfigModule } from '@nestjs/config';
 import { PatientsModule } from './patients/patients.module';
 import { PrismaService } from './prisma/prisma.service';
 import { AppointmentsModule } from './appointments/appointments.module';
+import { AuthModule } from './auth/auth.module';
+import { UsersModule } from './users/users.module';
 
 @Module({
   imports: [
     ConfigModule.forRoot({ isGlobal: true }),
+    AuthModule,
+    UsersModule,
     PatientsModule,
     AppointmentsModule,
   ],